v1.0
this capstone website ''Grandma's sweet' is a from scratch tutorial in two parts made by Mr. @john-smilga the coding addict on Youtube, It's all a bootstrap designed with one vanilla java script  feature of functionality, which i hope to keep working on in the coming days till it's all functional,the one feature is: when you click the cart icon on the sweet item: you can find above in the cart button in the top right added to the list and summed to the total.